updated directory structure and filter tests if no address provided
